refactor: Move ManagedAgentGraph alongside other managed types
