+++ packages/sdk/server-ai/src/api/graph/index.ts
@@ -2,4 +2,3 @@ export * from './types';
 export * from './LDGraphTracker';
 export * from './AgentGraphNode';
 export * from './AgentGraphDefinition';
-export * from './ManagedAgentGraph';
